+++ services_common/cpp/src/GuidUtil.cc
@@ -78,7 +78,7 @@ const char* GuidUtil::GetDeviceIdFileName()
             perror("lstat");
         }
         sFileName += buf;
-        unsigned found = sFileName.find_last_of("/");
+        size_t found = sFileName.find_last_of("/");
         if (found !=  std::string::npos) {
             sFileName.erase(found + 1);
         }
